+++ src/app/card-benefit/my-benefit.tsx
@@ -1,0 +1,3 @@
+export default function MyBenefit() {
+  return <div className="flex flex-col gap-5">내 카드</div>;
+}
